+++ src/lib/forms-schema/common.ts
@@ -99,4 +99,5 @@ export const JoiRange = Joi.extend((joi: typeof Joi) => ({
   ],
 }))
 
-export const base64DataRegex = /<[^>]*src="data:([a-zA-Z]*)\/([a-zA-Z]*);base64,([^"]*)".*>/m
+export const base64DataRegex =
+  /<[^>]*src="data:([a-zA-Z]*)\/([a-zA-Z]*);base64,([^"]*)".*>/m