+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `encryptPdf` property to `CP_HCMS` submission event type
+- `toDateDaysOffset`, `fromDateDaysOffset` and `defaultValueDaysOffset` properties to docs
 
 ## [0.9.11] - 2021-05-19